MDED-21 Fix after rebase on develop
